Move the date tests to be implementation specific
